# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateUserRequest:
    properties:
      name:
        maxLength: 64
        minLength: 1
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      role:
        enum:
        - admin
        - staff
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - name
    - password
    - role
    - username
    type: object
...
  models.DeleteUserRequest:
    properties:
      lid:
        type: string
    required:
    - lid
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  models.MessageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
...
  models.UpdateUserRequest:
    properties:
      lid:
        type: string
      name:
        maxLength: 64
        minLength: 1
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      role:
        enum:
        - admin
        - staff
        type: string
    required:
    - lid
    type: object
  models.UserResponse:
    properties:
      lid:
        type: string
      name:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: 這是一個整合 PostgreSQL 與 Gin 的設備管理系統。
  title: 設備管理系統 API
  version: "1.0"
paths:
...
  /api/private/user:
    delete:
      consumes:
      - application/json
      description: 刪除使用者 (僅限管理員)。不允許刪除自己或刪除最後一位管理員。
      parameters:
      - description: 刪除使用者請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DeleteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功刪除使用者
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: 不可刪除自己或最後一位管理員
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 使用者不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 刪除使用者
      tags:
      - private
    patch:
      consumes:
      - application/json
      description: 修改使用者資料 (僅限管理員)。不允許自降身分或移除最後一位管理員。
      parameters:
      - description: 修改使用者請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功修改使用者
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: 無權限修改自己角色或降級最後管理員
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 使用者不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 修改使用者資料
      tags:
      - private
    post:
      consumes:
      - application/json
      description: 新增一個使用者 (僅限管理員)
      parameters:
      - description: 新增使用者請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 成功新增使用者
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: 使用者名稱已存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 新增使用者
      tags:
      - private
  /api/private/users:
    get:
      consumes:
      - application/json
      description: 取得所有使用者資料，支援分頁 (僅限管理員)
      parameters:
      - description: 每頁筆數 (預設 50，最大 100)
        in: query
        name: limit
        type: integer
      - description: 偏移量 (預設 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 使用者列表
          headers:
            X-Total-Count:
              description: 符合條件的總筆數
              type: integer
          schema:
            items:
              $ref: '#/definitions/models.UserResponse'
            type: array
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 查看使用者列表
      tags:
      - private
...
securityDefinitions:
  BearerAuth:
    description: 請輸入 "Bearer <Your_JWT_Token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"